refactor: split Shopify index.ts into focused modules
Breaks the 501-line monolith into 7 focused files:
- client.ts: shopifyFetch, removeEdgesAndNodes, config constants
- cart-operations.ts: cart CRUD (createCart, addToCart, etc.)
- products.ts: product queries and reshape helpers
- collections.ts: collection queries and reshape helpers
- pages.ts: page queries
- menu.ts: menu query
- webhooks.ts: revalidation webhook handler

Barrel index.ts re-exports everything for backward compatibility.

Closes #105
